+++ src/main/java/com/mochimexa/ecommerce/controller/AddressController.java
@@ -2,9 +2,10 @@ package com.mochimexa.ecommerce.controller;
 
 import com.mochimexa.ecommerce.DTO.AddressRequestDTO;
 import com.mochimexa.ecommerce.DTO.AddressResponseDTO;
-import com.mochimexa.ecommerce.model.Address;
 import com.mochimexa.ecommerce.service.AddressService;
+
 import jakarta.validation.Valid;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,24 +22,32 @@ public class AddressController {
     }
 
     @GetMapping("/user/{userId}")
-    public List<AddressResponseDTO> getByUserId(@PathVariable Long userId) {
+    public List<AddressResponseDTO> getByUserId(
+            @PathVariable Integer userId
+    ) {
         return addressService.findByUserId(userId);
     }
 
     @PostMapping("/user/{userId}")
     @ResponseStatus(HttpStatus.CREATED)
-    public AddressResponseDTO create(@PathVariable Long userId, @Valid @RequestBody AddressRequestDTO dto) {
+    public AddressResponseDTO create(
+            @PathVariable Integer userId,
+            @Valid @RequestBody AddressRequestDTO dto
+    ) {
         return addressService.create(userId, dto);
     }
 
     @PutMapping("/{id}")
-    public AddressResponseDTO update(@PathVariable Long id, @Valid @RequestBody AddressRequestDTO dto) {
+    public AddressResponseDTO update(
+            @PathVariable Integer id,
+            @Valid @RequestBody AddressRequestDTO dto
+    ) {
         return addressService.update(id, dto);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Integer id) {
         addressService.deleteById(id);
     }
 }
