+++ src/main/java/com/mochimexa/ecommerce/controller/CartController.java
@@ -4,7 +4,9 @@ import com.mochimexa.ecommerce.DTO.AddToCartRequestDTO;
 import com.mochimexa.ecommerce.DTO.UpdateCartItemDTO;
 import com.mochimexa.ecommerce.model.Cart;
 import com.mochimexa.ecommerce.service.CartService;
+
 import jakarta.validation.Valid;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,27 +19,41 @@ public class CartController {
     public CartController(CartService cartService) {
         this.cartService = cartService;
     }
+
     @GetMapping("/user/{userId}")
-    public Cart getCartByUserId(@PathVariable Long userId) {
+    public Cart getCartByUserId(@PathVariable Integer userId) {
         return cartService.findByUserId(userId);
     }
+
     @PostMapping("/user/{userId}/items")
     @ResponseStatus(HttpStatus.CREATED)
-    public Cart addItem(@PathVariable Long userId, @Valid @RequestBody AddToCartRequestDTO dto) {
+    public Cart addItem(
+            @PathVariable Integer userId,
+            @Valid @RequestBody AddToCartRequestDTO dto
+    ) {
         return cartService.addItem(userId, dto);
     }
+
     @PutMapping("/items/{cartDetailId}")
-    public Cart updateItemQuantity(@PathVariable Long cartDetailId, @Valid @RequestBody UpdateCartItemDTO dto) {
-        return cartService.updateItemQuantity(cartDetailId, dto.getCantidad());
+    public Cart updateItemQuantity(
+            @PathVariable Integer cartDetailId,
+            @Valid @RequestBody UpdateCartItemDTO dto
+    ) {
+        return cartService.updateItemQuantity(
+                cartDetailId,
+                dto.getCantidad()
+        );
     }
+
     @DeleteMapping("/items/{cartDetailId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void removeItem(@PathVariable Long cartDetailId) {
+    public void removeItem(@PathVariable Integer cartDetailId) {
         cartService.removeItem(cartDetailId);
     }
+
     @DeleteMapping("/user/{userId}/clear")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void clearCart(@PathVariable Long userId) {
+    public void clearCart(@PathVariable Integer userId) {
         cartService.clearCart(userId);
     }
 }
