+++ src/main/java/com/mochimexa/ecommerce/repository/CategoryRepository.java
@@ -3,15 +3,16 @@ package com.mochimexa.ecommerce.repository;
 import com.mochimexa.ecommerce.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.Optional;
 import java.util.List;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 
-    //BUSCAR CATEGORIA POR NOMBRE
+    // BUSCAR CATEGORIA POR NOMBRE
     Optional<Category> findByNombreIgnoreCase(String nombre);
 
-    //LISTAR CATEGORIAS ACTIVAS
+    // LISTAR CATEGORIAS ACTIVAS
     List<Category> findByActivoTrue();
 }
