+++ src/main/java/com/mochimexa/ecommerce/repository/OrderDetailRepository.java
@@ -3,12 +3,12 @@ package com.mochimexa.ecommerce.repository;
 import com.mochimexa.ecommerce.model.OrderDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.List;
 
 @Repository
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long>{
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Integer> {
 
-    //DETALLES DE PEDIDO POR ID
-    List<OrderDetail> findByPedidoIdPedido(Long idPedido);
-
+    // DETALLES DE PEDIDO POR ID
+    List<OrderDetail> findByPedidoIdPedido(Integer idPedido);
 }
