+++ src/main/java/com/mochimexa/ecommerce/repository/PaymentRepository.java
@@ -3,13 +3,12 @@ package com.mochimexa.ecommerce.repository;
 import com.mochimexa.ecommerce.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.Optional;
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Long>{
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 
-    //BUSCAR PAGO POR ID PEDIDO
-    Optional<Payment> findByPedidoIdPedido(Long idPedido);
-
-
+    // BUSCAR PAGO POR ID PEDIDO
+    Optional<Payment> findByPedidoIdPedido(Integer idPedido);
 }
