+++ src/main/java/com/mochimexa/ecommerce/repository/PromotionRepository.java
@@ -3,11 +3,12 @@ package com.mochimexa.ecommerce.repository;
 import com.mochimexa.ecommerce.model.Promotion;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.List;
 
 @Repository
-public interface PromotionRepository extends JpaRepository<Promotion, Long>{
+public interface PromotionRepository extends JpaRepository<Promotion, Integer> {
 
-    //PROMOCIONES ACTIVAS
+    // PROMOCIONES ACTIVAS
     List<Promotion> findByActivoTrue();
 }
