+++ src/main/java/com/mochimexa/ecommerce/repository/UserRepository.java
@@ -1,27 +1,28 @@
 package com.mochimexa.ecommerce.repository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.mochimexa.ecommerce.model.User;
+
 import java.util.Optional;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    //BUSCAR USUARIO POR ID
-    Optional<User> findByIdUsuario(Long idUsuario);
+    // BUSCAR USUARIO POR ID
+    Optional<User> findByIdUsuario(Integer idUsuario);
 
-    //BUSCAR USUARIO POR CORREO
+    // BUSCAR USUARIO POR CORREO
     Optional<User> findByCorreo(String correo);
 
-    //VERIFICAR SI EXISTE EL CORREO
-    boolean existsByCorreo (String correo);
+    // VERIFICAR SI EXISTE EL CORREO
+    boolean existsByCorreo(String correo);
 
-    //BUSCAR USUARIOS ACTIVOS
+    // BUSCAR USUARIOS ACTIVOS
     List<User> findByActivo(boolean activo);
 
-    //BUSCAR USUARIOS POR ID DE SU ROL
-    List<User> findByRolIdRol(Long idRol);
-
+    // BUSCAR USUARIOS POR ID DE SU ROL
+    List<User> findByRolIdRol(Integer idRol);
 }
